training-data-service/requirements.txt
Flask
flask_sqlalchemy
psycopg2-binary
python-dotenv
werkzeug
numpy
openai
google-api-python-client
yt-dlp
openai-whisper
gunicorn
langchain
